Revert "Rename components"
This reverts commit 664ee5f4bbcbcba41b83cbcbdebbdf55c24cde0e.
