+++ components/blog/Container.tsx
@@ -1,7 +1,7 @@
 import Head from 'next/head';
 import { useRouter } from 'next/router';
 
-import Layout from '@/components/Layout';
+import Layout from '@/components/layout';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export default function Container(props: any) {
