+++ components/now-playing.tsx
