+++ pages/404.tsx
@@ -2,7 +2,7 @@ import Link from 'next/link';
 
 import useTranslation from '@/lib/useTranslation';
 
-import Layout from '@/components/Layout';
+import Layout from '@/components/layout';
 
 export default function NotFound(): JSX.Element {
   const { t } = useTranslation();
